requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pdf2image>=1.16.3
openai>=1.50.0
python-dotenv>=1.0.0
pillow>=10.1.0
PyPDF2>=3.0.1
requests>=2.31.0
httpx>=0.24.0